Intermediate UI Changes & Splash Screen Addition
- Additional changes have been made to the app's UI, and a splash screen has been added to the same.

- Flash have been created as needed for the same.
