+++ android/app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
    <application
         android:label="lg_gesture_and_voice_control"
         android:name="${applicationName}"
-        android:icon="@mipmap/ic_launcher">
+        android:icon="@mipmap/launcher_icon">
         <activity
             android:name=".MainActivity"
             android:exported="true"
